move internal code to /internal folder
