+++ internal/client.go
@@ -1,4 +1,4 @@
-package main
+package internal
 
 import (
 	"github.com/ethereum/go-ethereum/rpc"
@@ -8,7 +8,7 @@ type StatusGoClient struct {
 	rpcClient *rpc.Client
 }
 
-func newClient(url string) (*StatusGoClient, error) {
+func NewClient(url string) (*StatusGoClient, error) {
 	rpcClient, err := rpc.Dial(url)
 	if err != nil {
 		return nil, err
