+++ internal/json.go
@@ -1,4 +1,4 @@
-package main
+package internal
 
 import "fmt"
