+++ internal/loop.go
@@ -1,8 +1,13 @@
-package main
+package internal
 
 import (
 	"time"
+
+	"github.com/jroimartin/gocui"
 )
+
+// Rhread ending channel
+var threadDone = make(chan struct{})
 
 func FetchLoop(s *StateController, interval int) {
 	// Get the first peers fetch going sooner
@@ -16,4 +21,8 @@ func FetchLoop(s *StateController, interval int) {
 			s.Fetch()
 		}
 	}
+}
+func QuitLoop(g *gocui.Gui, v *gocui.View) error {
+	close(threadDone)
+	return gocui.ErrQuit
 }
