+++ internal/state.go
@@ -1,4 +1,4 @@
-package main
+package internal
 
 import (
 	"github.com/dannypsnl/redux/v2/rematch"
